requirements.txt
chromadb==0.5.0
sentence-transformers==3.0.1
transformers==4.40.2
torch>=2.2.0
pillow==10.3.0
pypdf==4.2.0
pyyaml==6.0.1
